Resolve warning

+++ src/main.cpp
@@ -8,7 +8,7 @@
 #include "allocatorSim.h"
 #include "allocatorMgr.h"
 
-typedef std::vector<std::tuple<size_t, size_t, size_t>> blockVector_t;
+typedef std::vector<std::tuple<uint64_t, uint64_t, size_t>> blockVector_t;
 
 std::vector<size_t> split_line(std::string str, const std::string c) {
     std::vector<size_t> vec;
@@ -30,7 +30,7 @@ std::vector<size_t> split_line(std::string str, const std::string c) {
     return vec;
 }
 
-std::tuple<size_t, size_t> process_trace(std::string filename, blockVector_t& block_list) {
+std::tuple<uint64_t, uint64_t> process_trace(std::string filename, blockVector_t& block_list) {
     std::ifstream file;
     file.open(filename);
     std::string line;
@@ -39,10 +39,10 @@ std::tuple<size_t, size_t> process_trace(std::string filename, blockVector_t& bl
         block_list.push_back(std::make_tuple(vec[0], vec[1], vec[2]));
     }
 
-    size_t min = std::numeric_limits<size_t>::max();
-    size_t max = 0;
+    uint64_t min = std::numeric_limits<uint64_t>::max();
+    uint64_t max = 0;
     std::for_each(block_list.begin(), block_list.end(),
-    [&min, &max](std::tuple<size_t, size_t, size_t> t) {
+    [&min, &max](std::tuple<uint64_t, uint64_t, size_t> t) {
         if (std::get<0>(t) < min) {
             min = std::get<0>(t);
         }
@@ -61,11 +61,11 @@ std::tuple<size_t, size_t> process_trace(std::string filename, blockVector_t& bl
     return std::make_tuple(min, max);
 }
 
-void run_allocator(const blockVector_t& block_list, const size_t min, const size_t max) {
+void run_allocator(const blockVector_t& block_list, const uint64_t min, const uint64_t max) {
     allocatorMgr alloc_mgr;
 
-    size_t count = 0;
-    for (int i = min; i <= max; i++) {
+    uint64_t count = 0;
+    for (uint64_t i = min; i <= max; i++) {
         if (std::get<0>(block_list[count]) == i) {
             auto ref = std::get<1>(block_list[count]) - std::get<0>(block_list[count]);
             alloc_mgr.malloc_block(std::get<2>(block_list[count]), ref);
@@ -81,7 +81,7 @@ int main() {
     allocSim.test_allocator();
 
     blockVector_t input_block_list;
-    size_t min, max;
+    uint64_t min, max;
     std::tie(min, max) = process_trace("./input/small_block_test.log", input_block_list);
     run_allocator(input_block_list, min, max);
